+++ src/main/java/com/phantomwing/thesilverage/tags/ModTags.java
@@ -3,7 +3,6 @@ package com.phantomwing.thesilverage.tags;
 import com.phantomwing.thesilverage.TheSilverAge;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.EntityTypeTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;